# doc/api.yaml
openapi: 3.0.0
info:
  title: Chat Application API
  description: Chat application REST API documentation
  version: 1.0.0
  contact:
    name: Chat App Team
    email: support@chatapp.com
    url: https://chatapp.com/support
servers:
- url: /api/v1
  description: API foundation path
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT authentication using the Bearer scheme.
        Example: "Authorization: Bearer {token}"
  responses:
    BadRequest:
      description: >
        The request was malformed or contained invalid parameters.
        Check the error details for specific validation issues.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 400
            message: "Invalid request parameters"
            trace: "req-123456"
            error: "validation.failed"
            errors:
              - "username: Must be between 3-50 characters"
              - "email: Invalid email format"
    Unauthorized:
      description: >
        Authentication failed or user lacks proper permissions.
        Include a valid JWT token in the Authorization header.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 401
            message: "Unauthorized: Invalid or expired token"
            trace: "req-789012"
            error: "auth.invalid_token"
            errors: []      
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 404
            message: "User not found"
            trace: "req-345678"
            error: "resource.not_found"
            errors: []     
    Forbidden:
      description: >
        Authenticated user does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    User:
      type: object
      description: User information model containing all user-related data
      required:
      - id
      - username
      - avatar_url
      - email
      - gender
      properties:
        id:
          type: string
          description: Unique identifier for the user
          pattern: ^[a-zA-Z0-9-]+$
          minLength: 1
          maxLength: 64
        username:
          type: string
          description: User's display name
          minLength: 1
          maxLength: 50
          pattern: ^[a-zA-Z0-9_-]+$
        name:
          type: string
          description: User's full name
          minLength: 1
          maxLength: 100
          pattern: ^[a-zA-Z0-9\s'-]+$
        email:
          type: string
          description: User's email address
          format: email
          minLength: 5
          maxLength: 255
          pattern: ^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$
        avatar_url:
          type: string
          description: URL of user's avatar
          format: url
          maxLength: 255
          minLength: 1
          pattern: ^https?://[^\s/$.?#].[^\s]*$
        gender:
          type: string
          description: User's gender
          enum:
          - male
          - female
          - unspecified
          default: unspecified
    UserEnvelope:
      description: Single user envelope used in responses.
      allOf:
        - $ref: '#/components/schemas/BaseSuccessResponse'
        - type: object
          required: 
            - data
          properties:
            data:
              type: object
              description: Payload containing a single user resource
              required: 
                - user
              properties:
                user:
                  $ref: '#/components/schemas/User'
...
    Message:
      type: object
      description: Message information model for chat messages
      required:
      - id
      - content
      - sender_id
      - conversation_id
      - created_at
      - type
      - status
      properties:
        id:
          type: string
          description: Unique identifier for the message
          pattern: ^[a-zA-Z0-9-]+$
          minLength: 1
          maxLength: 64
        content:
          type: string
          description: Message content text
          minLength: 1
          maxLength: 1000
          pattern: ^[\s\S]{1,1000}$
        sender_id:
          type: string
          description: ID of the message sender
          pattern: ^[a-zA-Z0-9-]+$
          minLength: 1
          maxLength: 64
        conversation_id:
          type: string
          description: ID of the present conversation
          pattern: ^[a-zA-Z0-9-]+$
          minLength: 1
          maxLength: 64
        created_at:
          type: string
          description:  Message creation timestamp in ISO 8601 format
          format: date-time
          minLength: 1
          maxLength: 64
          pattern: ^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d+)?(Z|[+-]\d{2}:\d{2})$
        type:
          type: string
          description: Type of message content
          enum:
          - text
          - image
          - video
          - file
        status:
          type: string
          description: Delivery status of the message
          enum:
          - sending
          - sent
          - delivered
          - read
          - failed
    MessageResourceEnvelopeCollection:
      description: Envelope containing a list of message resources and optional paging info.
      allOf:
        - $ref: '#/components/schemas/BaseSuccessResponse'
        - type: object
          description: Success response with a paginated messages collection
          required: 
            - data
          properties:
            data:
              $ref: '#/components/schemas/MessageCollection'
    MessageResourceEnvelope:
      description: Envelope containing a single message resource.
      allOf:
        - $ref: '#/components/schemas/BaseSuccessResponse'
        - type: object
          description: Success response with a single message resource
          required: 
            - data
          properties:
            data:
              type: object
              description: Payload containing the message resource
              required: 
                - resource
              properties:
                resource:
                  $ref: '#/components/schemas/Message'
    Error:
      type: object
      description: Standard error response model for API error conditions
      required:
      - code
      - message
      - trace
      - error
      - errors
      properties:
        code:
          type: integer
          format: int64
          description: HTTP error code
          minimum: 400
          maximum: 599
        message:
          type: string
          description: Human-readable error message
          minLength: 1
          maxLength: 255
          pattern: ^[\s\S]{1,255}$
        trace:
          type: string
          description: Error trace ID for debugging purposes
          pattern: ^[a-zA-Z0-9\s.,!?-]+$
          minLength: 1
          maxLength: 1000
        error:
          type: string
          description: Machine-readable error key (stable identifier for programmatic handling)
          minLength: 1
          maxLength: 100
          pattern: ^[a-z0-9_.:-]+$
          example: auth.invalid_token
        errors:
          type: array
          description: Collection of detailed error messages
          items:
            type: string
            minLength: 1
            maxLength: 255
            pattern: ^[\s\S]{1,255}$
          minItems: 0
          maxItems: 100
    BaseSuccessResponse:
      type: object
      description:  Base structure for all success responses
      required:
      - code
      - message
      properties:
        code:
          type: integer
          format: int64
          description: HTTP status code indicating success
          enum:
          - 200
          - 201
          minimum: 200
          maximum: 201
          example: 200
        message:
          type: string
          description: Human-readable success message
          minLength: 1
          maxLength: 255
          pattern: ^[\s\S]{1,1000}$
          example: "Operation succeeded"  
...
    MessageCollection:
      type: object
      description: Paginated collection of messages with metadata
      required:
        - messages
        - pagination
      properties:
        messages:
          type: array
          description: Array of message objects
          items:
            $ref: '#/components/schemas/Message'
          minItems: 1
          maxItems: 100
        pagination:
          type: object
          description: Pagination metadata for the messages page
          properties:
            total:
              type: integer
              format: int32
              description: Total number of messages available
              minimum: 0
              maximum: 2147483647 
            limit:
              type: integer
              format: int32
              description: Maximum number of messages returned per page
              minimum: 1
              maximum: 100
            offset:
              type: integer
              format: int32
              description: Starting position of the returned messages
              minimum: 0
...
  parameters:
    user_id:
      name: user_id
      in: path
      required: true
      description: ID of the user
      schema:
        type: string
        pattern: ^[a-zA-Z0-9-]+$
        minLength: 1
        maxLength: 64
        description: unique identifier of user
    id:
      name: id
      in: path
      required: true
      description: ID of the resource
      schema:
        type: string
        pattern: ^[a-zA-Z0-9-]+$
        minLength: 1
        maxLength: 64
        description: uniquie identifier
paths:
...
  /users/me:
    get:
      tags:
      - users
      operationId: getUserInfo
      summary: Get current user info
      description: Retrieve current user's profile information
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
              example:
                code: 200
                message: User information retrieved
                data:
                  user:
                    id: usr-123456
                    username: john_doe
                    name: John Doe
                    email: john@example.com
                    avatar_url: https://example.com/avatars/john_doe.jpg
                    gender: male
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /users/profile/{user_id}:
    parameters:
      - $ref: '#/components/parameters/user_id'
    get:
      tags:
      - users
      operationId: getUserProfile
      summary: Get user profile
      description: Retrieve detailed profile information for a specific user
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
              example:
                code: 200
                message: User information retrieved
                data:
                  user:
                    id: usr-123456
                    username: john_doe
                    name: John Doe
                    email: john@example.com
                    avatar_url: https://example.com/avatars/john_doe.jpg
                    gender: male
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /messages:
    get:
      tags:
        - messages
      operationId: getConversation
      summary: Get conversation messages
      description: Retrieve messages from a specific conversation
      security:
        - BearerAuth: []
      parameters:
        - name: conversation_id
          in: query
          required: true
          description: ID of the conversation
          schema:
            type: string
            pattern: ^[a-zA-Z0-9-]+$
            description: Unique identifier of the conversation to retrieve messages from
            minLength: 1
            maxLength: 64
        - name: limit
          in: query
          required: false
          description: Maximum number of messages to return
          schema:
            type: integer
            format: int64
            minimum: 1
            maximum: 100
            default: 20
            description: Maximum number of messages to return in a single page
        - name: before
          in: query
          description: Get messages before this timestamp (ISO 8601 format)
          schema:
            type: string
            description: ISO 8601 timestamp to get messages before this time
            format: date-time
            minLength: 20
            maxLength: 35
      responses:
        '200':
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResourceEnvelopeCollection'
              example:
                code: 200
                message: Messages retrieved successfully
                data:
                  messages:
                    - id: '1'
                      content: Hello!
                      sender_id: '1'
                      conversation_id: 'c-1'
                      created_at: '2023-10-01T12:00:00Z'
                      type: text
                      status: delivered
                  pagination:
                    total: 120
                    limit: 20
                    offset: 0
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    
    post:
      tags:
      - messages
      operationId: sendMessage
      summary: Send a message
      description: Send a new message to a user or group
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Send a message
              required:
              - conversation_id
              - content
              properties:
                conversation_id:
                  type: string
                  description: ID of the target conversation
                  pattern: ^[a-zA-Z0-9-]+$
                  minLength: 1 
                  maxLength: 64
                content:
                  type: string
                  description: Message content
                  minLength: 1
                  maxLength: 1000
                  pattern: '^[\s\S]{1,1000}$'
                type:
                  type: string
                  description: Message type
                  enum: 
                  - text
                  - image
                  - video
                  - file
                  default: text
      responses:
        '201':
          description: Message sent successfully
          content:
            application/json:
              schema:
               $ref: '#/components/schemas/MessageResourceEnvelope'
              example:
                code: 201
                message: Message sent successfully
                data:
                  resource:
                    id: '123'
                    content: Hello!
                    sender_id: '1'
                    conversation_id: 'c-1'
                    created_at: '2023-10-01T12:00:00Z'
                    type: text
                    status: sent
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /messages/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags:
      - messages
      operationId: getMessageById
      summary: Get a specific message
      description: Retrieve details of a message by its ID
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Message details retrieved successfully
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/MessageResourceEnvelope'
              example:
                code: 200
                message: Message details retrieved
                data:
                  resource:
                    id: '123'
                    content: Hello!
                    sender_id: '1'
                    conversation_id: 'c-1'
                    created_at: '2023-10-01T12:00:00Z'
                    type: text
                    status: delivered
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - messages
      operationId: deleteMessage
      summary: Delete a message
      description: Permanently delete a message
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Message deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseSuccessResponse'
              example:
                code: 200
                message: Message deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /messages/{id}/forward:
    parameters:
      - $ref: '#/components/parameters/id'
    post:
      tags:
      - messages
      operationId: forwardMessage
      summary: Forward a message
      description: Forward an existing message to another user or group
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Forward a message to another conversation
              required:
              - conversation_id
              properties:
                conversation_id:
                  type: string
                  description: Target conversation ID
                  pattern: '^[a-zA-Z0-9-]+$'
                  minLength: 1
                  maxLength: 64
      responses:
        '200':
          description: Message forwarded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseSuccessResponse'
              example:
                code: 200
                message: Message forwarded successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /messages/{id}/uncomment:
    post:
      tags:
      - messages
      operationId: uncommentMessage
      summary: Remove a comment from a message
      description: Delete a comment from a message
      security:
      - BearerAuth: []
      parameters:
      - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Comment removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseSuccessResponse'
              example:
                code: 200
                message: Comment removed successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...